+++ CHANGELOG.md
@@ -8,6 +8,8 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - xxxx-xx-xx
 
+## [v1.17.0] - 2022-05-10
+
 ### Added
 
 - New configuration option: `assetCdnBaseUrl` [#159](https://github.com/sharetribe/flex-sdk-js/pull/159)
@@ -16,7 +18,6 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Request Assets in JSON format instead of Transit [#160](https://github.com/sharetribe/flex-sdk-js/pull/160)
 - Read response data as Transit only if Content-Type header is `application/transit+json` [#161](https://github.com/sharetribe/flex-sdk-js/pull/161)
-
 
 ## [v1.16.0] - 2022-03-23